requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
playwright>=1.40.0
